Created a RandomModule for NumberCards
